COS30017assignment2
Created app our images of reasonable size that describe locations near you. You can use any images if the images are not objectionable.
When an image is clicked, a new screen with further details needs to be shown. The data needs to be passed as a Parcelable object. The non-TextView widget needs to show the correct value on activity creation.
The title and rating should be shown under each image in the first activity. At least one of your attributes should use a non-TextView widget, such as a RatingBar, a Switch, or a RadioGroup.
The data kept for each image should include the name of the location, a city/state/country to help place the location, the date of your last visit, and a rating from 0–5.
A model for the image data is used, which implements the Parcelable protocol, and is populated. Image data should not be stored on a disk, but rather kept in memory.
A grid, Constraint or Linear layout, in which case TableLayout might be useful.
Use of at least two styles used in multiple locations.
Code is commented, indented correctly, uses understandable variable names, shows some evidence of decomposition and design to avoid repeated code, and makes use of Android practices. For example:
string resources instead of hard-coded strings
Kotlin features like scope functions.